refactor(runner): split runner.py into resources/containers/params modules
Behavior-preserving code-organization change. runner.py (858 lines) mixed
five concerns; the cohesive, self-contained pieces move into new modules
inside the qsm_ci package, leaving runner.py as CLI orchestration.

- qsm_ci/resources.py: _ResourceSampler + _parse_mem_usage / _parse_cpu_perc
  / _parse_bytes / _MEM_UNITS (the docker-stats resource sampler). The
  _stop_event attribute name is preserved exactly (renaming it to _stop
  shadows threading.Thread._stop and DNF'd every sampled run).
- qsm_ci/containers.py: container execution (_run_container) and the
  runner/image helpers it uses (RUNNERS, _OCI_ENGINES, _have, check_runner,
  check_docker, _build_oci, _apptainer_image, _param_env). Keeps the
  apptainer network-isolation caveat comment.
- qsm_ci/params.py: params.json / sidecar / echo assembly (_params_dict,
  _params_summary, _place_echoes, _place_input, _echo_key, _nifti_voxel_size,
  _ensure_te, _looks_like_sidecar, _sidecar_te, _sidecar_to_params) plus the
  OPTIONAL_ARTIFACTS / STACKABLE_ARTIFACTS constants.

No logic, messages, or CLI behavior changed. runner.py imports what it now
needs from the new modules and re-exports check_docker, RUNNERS,
OPTIONAL_ARTIFACTS, STACKABLE_ARTIFACTS and _ResourceSampler so existing
importers (qsm_ci.cli, the sampler regression test, scripts/pipeline docs)
keep working.

Adds the resource-sampler regression test (was on a separate branch) that
pins the _stop -> _stop_event fix.
