+++ qsm_ci/runner.py
@@ -10,29 +10,19 @@ from __future__ import annotations
 
 import argparse
 import difflib
-import glob
-import gzip
 import json
 import math
 import os
-import re
 import shutil
-import struct
-import subprocess
-import threading
-import time
 from pathlib import Path
-from uuid import uuid4
 
+from .containers import (RUNNERS, _run_container, check_docker,  # noqa: F401 — check_docker re-exported for back-compat
+                         check_runner)
+from .resources import _ResourceSampler  # noqa: F401 — re-exported for the sampler regression test
+from .params import (OPTIONAL_ARTIFACTS, STACKABLE_ARTIFACTS, _looks_like_sidecar,
+                     _params_dict, _params_summary, _place_echoes,
+                     _place_input, _sidecar_to_params)
 from .stages import ARTIFACT_FILE, ARTIFACT_KIND, STAGES
-
-# Consumed artifacts that aren't required to run a stage (only some methods use them, e.g. MEDI
-# uses magnitude; plain TKD does not). Everything else the stage consumes is required.
-OPTIONAL_ARTIFACTS = {"magnitude"}
-
-# Multi-echo artifacts: a stage wants one 4D NIfTI (x,y,z,echo), but a caller with BIDS data has one
-# 3D file per echo. These flags accept several files and we stack them into the 4D artifact.
-STACKABLE_ARTIFACTS = {"phase", "magnitude"}
 
 
 def _consumes(algo: dict) -> list:
@@ -162,158 +152,6 @@ def _closest_slugs(target: str) -> "list[str]":
     return difflib.get_close_matches(target, names, n=3, cutoff=0.4)
 
 
-def _nifti_voxel_size(path) -> "list[float] | None":
-    """Read pixdim[1:4] (mm) straight from a NIfTI-1 header — no nibabel dependency."""
-    if not path or not Path(path).exists():
-        return None
-    opener = gzip.open if str(path).endswith(".gz") else open
-    try:
-        with opener(path, "rb") as f:
-            hdr = f.read(352)
-        if len(hdr) < 352:
-            return None
-        for endian in ("<", ">"):  # header endianness is whichever makes sizeof_hdr == 348
-            if struct.unpack(endian + "i", hdr[0:4])[0] == 348:
-                pixdim = struct.unpack(endian + "8f", hdr[76:108])
-                vs = [abs(pixdim[1]), abs(pixdim[2]), abs(pixdim[3])]
-                return vs if all(v > 0 for v in vs) else None
-    except Exception:  # noqa: BLE001 — best-effort; caller falls back to a default
-        return None
-    return None
-
-
-def _place_input(src, dest: Path) -> None:
-    """Put a consumed NIfTI at <name>.nii.gz, gzip-compressing a plain .nii on the way in."""
-    src = str(src)
-    if src.endswith(".nii") and str(dest).endswith(".nii.gz"):
-        with open(src, "rb") as fi, gzip.open(dest, "wb") as fo:
-            shutil.copyfileobj(fi, fo)
-    else:
-        shutil.copy(src, dest)
-
-
-def _echo_key(path) -> int:
-    m = re.search(r"echo-?(\d+)", str(path))
-    return int(m.group(1)) if m else 0
-
-
-def _place_echoes(paths: list, dest: Path, log) -> None:
-    """Place a multi-echo artifact: one file goes in as-is; several 3D echoes are stacked into 4D.
-
-    When every filename carries a BIDS `echo-<n>`, echoes are ordered by that number (so echo-10
-    sorts after echo-2); otherwise the given order is kept. The stacked file must line up with the
-    `TE` list in params.json."""
-    if len(paths) == 1:
-        _place_input(paths[0], dest)
-        return
-    import nibabel as nib
-    import numpy as np
-    ordered = sorted(paths, key=_echo_key) if all(re.search(r"echo-?\d+", str(p)) for p in paths) else list(paths)
-    imgs = [nib.load(str(p)) for p in ordered]
-    data = np.stack([im.get_fdata(dtype=np.float64) for im in imgs], axis=-1)
-    dest.parent.mkdir(parents=True, exist_ok=True)
-    nib.save(nib.Nifti1Image(data.astype(np.float32), imgs[0].affine), str(dest))
-    log(f"  stacked {len(ordered)} echoes → {dest.name}")
-
-
-def _ensure_te(te: list, stage: str) -> list:
-    """BFR/dipole don't use TE physically — the field is already in ppm and the TE·B0 scaling that
-    some recon code applies cancels out — but that code still indexes TE(1). Supply a nominal
-    placeholder so an omitted --te writes TE=[1.0] rather than an empty list that blows up the
-    container (MATLAB: "Index exceeds array bounds"). Never fabricates TE for phase-consuming
-    stages (field-mapping / end-to-end) — those genuinely need a real echo time."""
-    if te or "phase" in STAGES[stage]["consumes"]:
-        return te
-    return [1.0]
-
-
-def _params_dict(args, stage: str) -> dict:
-    """Assemble a params.json dict from --te/--field-strength/--b0-dir/--voxel-size (+ defaults).
-
-    B0_dir defaults to +z; voxel size is read from the primary input's NIfTI header when not given.
-    Echo times and field strength are only required for stages that consume phase (field-mapping);
-    BFR/dipole work in ppm and don't use them, so they get harmless placeholders.
-    """
-    consumes = STAGES[stage]["consumes"]
-    needs_echo = "phase" in consumes
-    te = list(args.te) if args.te else []
-    b0 = args.field_strength
-    if needs_echo and (not te or b0 is None):
-        raise SystemExit(
-            f"the {stage} stage needs echo times and field strength — pass "
-            "--te SEC [SEC ...] and --field-strength TESLA, or give a --params file.")
-    te = _ensure_te(te, stage)  # BFR/dipole: nominal placeholder if omitted (cancels; never empty)
-    if b0 is None:
-        b0 = 3.0  # unused by BFR/dipole; a contract placeholder
-    b0_dir = list(args.b0_dir) if args.b0_dir is not None else [0.0, 0.0, 1.0]
-    if args.voxel_size is not None:
-        voxel = list(args.voxel_size)
-    else:
-        primary = getattr(args, consumes[0], None)
-        voxel = _nifti_voxel_size(primary) or [1.0, 1.0, 1.0]
-    return {"TE": [float(t) for t in te], "B0": float(b0),
-            "B0_dir": [float(x) for x in b0_dir], "voxel_size": [float(v) for v in voxel]}
-
-
-def _params_summary(params: dict, stage: str) -> str:
-    """One-line echo of the params actually used by a stage. Field-mapping consumes TE + B0;
-    BFR/dipole work in ppm and use only B0 direction + voxel size, so don't imply otherwise."""
-    if "phase" in STAGES[stage]["consumes"]:
-        return (f"TE={params['TE']} B0={params['B0']} "
-                f"B0_dir={params['B0_dir']} voxel_size={params['voxel_size']}")
-    return (f"B0_dir={params['B0_dir']} voxel_size={params['voxel_size']}  "
-            f"(TE / field strength not used by this stage)")
-
-
-def _looks_like_sidecar(obj: dict) -> bool:
-    """A BIDS MEGRE sidecar carries these keys; a QSM-CI params.json does not."""
-    return isinstance(obj, dict) and ("EchoTime" in obj or "MagneticFieldStrength" in obj)
-
-
-def _sidecar_te(sidecar_path: Path) -> "list[float]":
-    """Echo times (s) for the whole acquisition: all `*part-phase*MEGRE.json` in the sidecar's
-    directory, in echo order. Falls back to just this file's EchoTime."""
-    def echo_no(f: str) -> int:
-        m = re.search(r"echo-(\d+)", f)
-        return int(m.group(1)) if m else 0
-    tes = []
-    for f in sorted(glob.glob(str(sidecar_path.parent / "*part-phase*MEGRE.json")), key=echo_no):
-        try:
-            v = json.load(open(f)).get("EchoTime")
-        except Exception:  # noqa: BLE001
-            v = None
-        if v is not None:
-            tes.append(float(v))
-    if not tes:
-        v = json.load(open(sidecar_path)).get("EchoTime")
-        tes = [float(v)] if v is not None else []
-    return tes
-
-
-def _sidecar_to_params(path: Path, obj: dict, args, stage: str) -> dict:
-    """Map a BIDS MEGRE phase sidecar onto the QSM-CI params.json schema.
-
-    Voxel size is read from the input NIfTI header (the authoritative grid); the sidecar's
-    `VoxelSize` is only a fallback. Any explicit acquisition flag the user passed wins.
-    """
-    consumes = STAGES[stage]["consumes"]
-    primary = getattr(args, consumes[0], None)
-    te = _sidecar_te(path)
-    b0 = obj.get("MagneticFieldStrength", 3.0)
-    b0_dir = obj.get("B0_dir") or [0.0, 0.0, 1.0]
-    voxel = _nifti_voxel_size(primary) or obj.get("VoxelSize") or [1.0, 1.0, 1.0]
-    if args.te:
-        te = args.te
-    if args.field_strength is not None:
-        b0 = args.field_strength
-    if args.b0_dir is not None:
-        b0_dir = args.b0_dir
-    if args.voxel_size is not None:
-        voxel = args.voxel_size
-    te = _ensure_te(te, stage)  # BFR/dipole: nominal placeholder if the sidecar carried no TE
-    return {"TE": [float(t) for t in te], "B0": float(b0),
-            "B0_dir": [float(x) for x in b0_dir], "voxel_size": [float(v) for v in voxel]}
-
 
 def _inputs_summary(slug: str, algo: dict) -> str:
     """Tell the user exactly which inputs a valid slug's stage needs, with an example."""
@@ -361,258 +199,6 @@ def list_command(argv=None, log=print) -> int:
     """`qsm-ci list` — show the reference algorithms available to run."""
     log(_algorithms_help())
     return 0
-
-
-RUNNERS = ("docker", "podman", "apptainer", "local")
-_OCI_ENGINES = ("docker", "podman")  # daemonless podman is CLI-compatible with docker
-
-
-def _have(binary: str) -> bool:
-    return shutil.which(binary) is not None
-
-
-def check_runner(runner: str) -> bool:
-    """Is the tooling for this runner available?"""
-    if runner == "local":
-        return True
-    if runner == "docker":  # also confirm the daemon answers
-        try:
-            return subprocess.run(["docker", "version"], capture_output=True).returncode == 0
-        except FileNotFoundError:
-            return False
-    return _have(runner)
-
-
-def check_docker() -> bool:  # kept for back-compat
-    return check_runner("docker")
-
-
-def _build_oci(algo: dict, engine: str, log) -> str:
-    """docker/podman: PULL the submission's prebuilt image: and return its ref — never build.
-
-    CI must not build containers. Every submission publishes a prebuilt ``image:``; a folder
-    Dockerfile is only the build recipe (used out-of-band to produce that image), never built here.
-    Pull the tag; if the pull fails (offline / registry hiccup) fall back to a locally cached copy;
-    if neither is available, stop with a clear message telling the author to build and push first."""
-    tag = algo.get("image")
-    if not tag:
-        raise SystemExit(
-            "algorithm.yml has no image:. Publish a prebuilt image (build and push it first) and set "
-            "image: to its reference — the runner pulls images, it does not build containers.")
-    log(f"↓ pulling {tag}")
-    if subprocess.run([engine, "pull", tag], capture_output=True).returncode != 0:
-        # Offline / registry hiccup — fall back to a locally cached copy if one exists.
-        if subprocess.run([engine, "image", "inspect", tag], capture_output=True).returncode != 0:
-            raise SystemExit(
-                f"could not pull {tag} and no local copy is cached. Build and push the image first; "
-                "the runner does not build containers (a folder Dockerfile is only the build recipe).")
-        log(f"  (pull failed — using locally cached {tag})")
-    return tag
-
-
-def _apptainer_image(algo: dict) -> str:
-    """apptainer runs from a docker:// ref or a .sif — it can't build a Dockerfile itself.
-
-    A published ``image:`` is used as-is (pulled & converted on the fly) exactly like the docker
-    path does — a folder Dockerfile is only the build recipe, never built here. Only when no
-    ``image:`` is set does a local Dockerfile become a hard error (apptainer can't build it)."""
-    img = algo.get("image")
-    if img:
-        if "://" in img or img.endswith(".sif") or os.path.exists(img):
-            return img
-        return f"docker://{img}"  # plain registry ref -> pull & convert on the fly
-    if (algo["dir"] / "Dockerfile").exists():
-        raise SystemExit(
-            "apptainer can't build a Dockerfile. Build it first with --runner docker/podman, "
-            "or set image: to a prebuilt reference (docker://…, a registry ref, or a .sif).")
-    raise SystemExit("algorithm.yml has no image: for apptainer to run")
-
-
-def _param_env(input_dir: Path) -> "dict[str, str]":
-    """Layer B — expose params.json + config.json as ``QSMCI_*`` env vars so a run.sh needn't parse
-    JSON (no `jq` needed): ``QSMCI_B0``, ``QSMCI_TE`` (space-separated echoes), ``QSMCI_TE0`` (first
-    echo), ``QSMCI_B0_DIR``, ``QSMCI_VOXEL_SIZE``, and ``QSMCI_SET_<NAME>`` per --set override. The
-    JSON files are still written (Layer A), so this is purely additive."""
-    env: "dict[str, str]" = {}
-    pj = input_dir / ARTIFACT_FILE["params"]
-    if pj.exists():
-        try:
-            p = json.loads(pj.read_text())
-        except Exception:  # noqa: BLE001
-            p = {}
-        te = p.get("TE") or []
-        if te:
-            env["QSMCI_TE"] = " ".join(str(t) for t in te)
-            env["QSMCI_TE0"] = str(te[0])
-        if p.get("B0") is not None:
-            env["QSMCI_B0"] = str(p["B0"])
-        if p.get("B0_dir"):
-            env["QSMCI_B0_DIR"] = " ".join(str(x) for x in p["B0_dir"])
-        if p.get("voxel_size"):
-            env["QSMCI_VOXEL_SIZE"] = " ".join(str(x) for x in p["voxel_size"])
-    cj = input_dir / "config.json"
-    if cj.exists():
-        try:
-            for k, v in json.loads(cj.read_text()).items():
-                env[f"QSMCI_SET_{str(k).upper()}"] = str(v)
-        except Exception:  # noqa: BLE001
-            pass
-    return env
-
-
-_MEM_UNITS = {"B": 1, "KB": 1e3, "MB": 1e6, "GB": 1e9, "TB": 1e12,
-              "KIB": 1024, "MIB": 1024 ** 2, "GIB": 1024 ** 3, "TIB": 1024 ** 4}
-
-
-def _parse_bytes(tok: str) -> "float | None":
-    """"1.2GiB" / "512MiB" / "0B" -> bytes. None if unparseable (docker sometimes emits "--")."""
-    tok = tok.strip()
-    m = re.match(r"^([0-9.]+)\s*([A-Za-z]+)$", tok)
-    if not m:
-        return None
-    val, unit = m.group(1), m.group(2).upper()
-    try:
-        return float(val) * _MEM_UNITS.get(unit, 1)
-    except ValueError:
-        return None
-
-
-def _parse_mem_usage(field: str) -> "float | None":
-    """docker/podman `stats` MemUsage is "<used> / <limit>"; we want the used side in bytes."""
-    used = field.split("/")[0]
-    return _parse_bytes(used)
-
-
-def _parse_cpu_perc(field: str) -> "float | None":
-    """"380.00%" -> 3.8 cores (CPUPerc is summed across cores, so >100% means multi-core)."""
-    field = field.strip().rstrip("%").strip()
-    try:
-        return float(field) / 100.0
-    except ValueError:
-        return None
-
-
-class _ResourceSampler(threading.Thread):
-    """Poll `<engine> stats <name>` on an interval while a container runs, recording a time series of
-    memory (bytes) and CPU (cores). A single bad/empty sample is skipped, never raised — the sampler
-    must not perturb the run it's measuring. On stop() the series is written to `out_path` as JSON."""
-
-    def __init__(self, name: str, engine: str, out_path: Path, interval: float = 1.0):
-        super().__init__(daemon=True)
-        self.name, self.engine, self.out_path, self.interval = name, engine, Path(out_path), interval
-        # NB: must NOT be named `_stop` — threading.Thread has a private `_stop()` method that join()
-        # calls internally; shadowing it with an Event breaks join() ("'Event' object is not callable")
-        # and made every sampled run crash -> DNF.
-        self._stop_event = threading.Event()
-        self.t, self.mem_bytes, self.cpu_cores = [], [], []
-
-    def _sample(self):
-        try:
-            r = subprocess.run(
-                [self.engine, "stats", self.name, "--no-stream",
-                 "--format", "{{.MemUsage}}|{{.CPUPerc}}"],
-                capture_output=True, text=True, timeout=5)
-        except Exception:  # noqa: BLE001 — engine hiccup / container not up yet; skip this tick
-            return
-        if r.returncode != 0:
-            return
-        line = (r.stdout or "").strip().splitlines()
-        if not line:
-            return
-        parts = line[0].split("|")
-        if len(parts) != 2:
-            return
-        mem = _parse_mem_usage(parts[0])
-        cpu = _parse_cpu_perc(parts[1])
-        if mem is None or cpu is None:
-            return
-        self.mem_bytes.append(mem)
-        self.cpu_cores.append(cpu)
-
-    def run(self):
-        t0 = time.time()
-        # Poll immediately, then every `interval` seconds until stopped — so even a very short run
-        # (0–2 samples) records something rather than nothing.
-        while not self._stop_event.is_set():
-            before = time.time()
-            self._sample()
-            if self.mem_bytes:  # only timestamp accepted samples, so t/mem/cpu stay aligned
-                if len(self.t) < len(self.mem_bytes):
-                    self.t.append(round(before - t0, 3))
-            elapsed = time.time() - before
-            self._stop_event.wait(max(0.0, self.interval - elapsed))
-
-    def stop(self):
-        self._stop_event.set()
-
-    def write(self):
-        try:
-            self.out_path.parent.mkdir(parents=True, exist_ok=True)
-            doc = {
-                "interval_s": self.interval,
-                "t": self.t,
-                "mem_bytes": self.mem_bytes,
-                "cpu_cores": self.cpu_cores,
-                "mem_peak_bytes": max(self.mem_bytes) if self.mem_bytes else 0,
-                "cpu_cores_max": max(self.cpu_cores) if self.cpu_cores else 0,
-                "sampler": "docker-stats",
-                "runner": self.engine,
-            }
-            self.out_path.write_text(json.dumps(doc))
-        except Exception:  # noqa: BLE001 — profiling is best-effort; never sink the run over it
-            pass
-
-
-def _run_container(algo, input_dir, output_dir, runner, log) -> float:
-    if output_dir.exists():
-        shutil.rmtree(output_dir)
-    output_dir.mkdir(parents=True)
-    penv = _param_env(input_dir)  # Layer B: acquisition params + overrides as QSMCI_* env vars
-    t0 = time.time()
-    if runner in _OCI_ENGINES:
-        image = _build_oci(algo, runner, log)
-        log(f"⚙ running container ({runner}: {image})")
-        # rootless podman: keep-id maps your host uid inside, so files written to the /output
-        # bind mount come back owned by you. docker (root daemon): run as your uid directly.
-        id_args = (["--userns=keep-id"] if runner == "podman"
-                   else ["--user", f"{os.getuid()}:{os.getgid()}"])
-        e_args = [a for k, v in penv.items() for a in ("-e", f"{k}={v}")]
-        # Name the container so a background sampler can poll `<engine> stats <name>` for a
-        # memory-over-time / CPU-over-time trace (opt-in via $QSMCI_RESOURCES_OUT).
-        name = f"qsm-ci-{uuid4().hex[:12]}"
-        res_out = os.environ.get("QSMCI_RESOURCES_OUT")
-        sampler = None
-        if res_out:
-            sampler = _ResourceSampler(name, runner, Path(res_out), interval=1.0)
-            sampler.start()
-        try:
-            subprocess.run([
-                runner, "run", "--rm", "--network", "none", "--name", name, *id_args, *e_args,
-                "-v", f"{algo['dir']}:/algo:ro",
-                "-v", f"{input_dir}:/input:ro", "-v", f"{output_dir}:/output",
-                image, "bash", "/algo/run.sh",
-            ], check=True)
-        finally:
-            if sampler is not None:
-                sampler.stop()
-                sampler.join(timeout=6)
-                sampler.write()
-    elif runner == "apptainer":
-        image = _apptainer_image(algo)
-        log(f"⚙ running container (apptainer: {image})")
-        log("  note: apptainer runs without enforced network isolation here; CI uses --network none.")
-        e_args = [a for k, v in penv.items() for a in ("--env", f"{k}={v}")]
-        subprocess.run([
-            "apptainer", "exec", "--no-home", "--cleanenv", *e_args,
-            "-B", f"{algo['dir']}:/algo:ro",
-            "-B", f"{input_dir}:/input:ro", "-B", f"{output_dir}:/output",
-            image, "bash", "/algo/run.sh",
-        ], check=True)
-    else:  # local
-        log("⚙ running run.sh directly (--runner local)")
-        subprocess.run(["bash", str(algo["dir"] / "run.sh"), str(input_dir), str(output_dir)],
-                       check=True, env={**os.environ, **penv})
-    return time.time() - t0
 
 
 def _score(recon: Path, artifact: str, truth: Path, mask: Path, seg: "Path | None") -> dict:
